requirements.txt
alembic==1.0.0
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Migrate==2.2.1
Flask-MongoAlchemy==0.7.2
flask-restplus==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10.1
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.4
mccabe==0.6.1
mongo==0.2.0
MongoAlchemy==0.19
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
pylint==2.1.1
pymongo==2.8.1
pytest==3.8.1
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
redis==2.10.6
requests==2.22.0
six==1.11.0
SQLAlchemy==1.3.4
typed-ast==1.1.0
urllib3==1.25.3
Werkzeug==0.14.1
wrapt==1.10.11
